Update package name from api to clients
