--- src/main/java/com/viepovsky/api/car/CarApiDto.java
@@ -1,16 +1,0 @@
-package com.viepovsky.api.car;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-class CarApiDto {
-    @JsonProperty("model")
-    private String model;
-}
