--- src/main/java/com/viepovsky/api/weather/dto/LocationsDto.java
@@ -1,16 +1,0 @@
-package com.viepovsky.api.weather.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class LocationsDto {
-    @JsonProperty("id")
-    private int cityId;
-}
