+++ src/main/java/com/viepovsky/clients/car/CarApiDto.java
@@ -1,0 +1,16 @@
+package com.viepovsky.clients.car;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+class CarApiDto {
+    @JsonProperty("model")
+    private String model;
+}
