+++ src/main/java/com/viepovsky/clients/weather/StoredForecastRepository.java
