+++ src/main/java/com/viepovsky/clients/weather/WeatherApiConfig.java
@@ -1,0 +1,16 @@
+package com.viepovsky.clients.weather;
+
+import lombok.Getter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Getter
+@Component
+class WeatherApiConfig {
+    @Value("${weather.api.endpoint}")
+    private String weatherApiEndpoint;
+    @Value("${weather.api.key}")
+    private String weatherApiKey;
+    @Value("${weather.api.host}")
+    private String weatherApiHost;
+}
