+++ src/main/java/com/viepovsky/clients/weather/dto/LocationDto.java
@@ -1,0 +1,18 @@
+package com.viepovsky.clients.weather.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class LocationDto {
+    @JsonProperty("locations")
+    private List<LocationsDto> locations;
+}
