+++ src/main/java/com/viepovsky/clients/weather/dto/LocationsDto.java
@@ -1,0 +1,16 @@
+package com.viepovsky.clients.weather.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class LocationsDto {
+    @JsonProperty("id")
+    private int cityId;
+}
